requirements.txt
discord.py==2.4.0
nodriver==0.38.post1
PyNaCl==1.5.0
python-dotenv==1.0.1
selenium==4.27.1
undetected_chromedriver==3.5.5
win_inet_pton==1.1.0
